Switched back to robotics version.

+++ app/src/main/java/com/frizzl/app/frizzleapp/lesson/exercise/MultipleResponse.java
@@ -86,7 +86,7 @@ public class MultipleResponse extends Exercise {
                 }
             });
 
-            TableRow row = rows.get((int)i / COLS_NUM);
+            TableRow row = rows.get(i / COLS_NUM);
             row.addView(button);
         }
     }
